Implement gap controller
